+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[8.1].define(version: 2026_05_25_160070) do
+ActiveRecord::Schema[8.1].define(version: 2026_05_25_160100) do
   # These are extensions that must be enabled in order to support this database
   enable_extension "pg_catalog.plpgsql"
 
@@ -75,13 +75,18 @@ ActiveRecord::Schema[8.1].define(version: 2026_05_25_160070) do
   end
 
   create_table "printers", force: :cascade do |t|
+    t.decimal "ambient_temp", precision: 6, scale: 2
     t.string "camera_url"
     t.datetime "created_at", null: false
+    t.decimal "enclosure_humidity", precision: 6, scale: 2
+    t.decimal "enclosure_temp", precision: 6, scale: 2
+    t.datetime "environment_updated_at"
     t.string "ha_base_sensor"
     t.string "hostname", null: false
     t.string "location"
     t.string "model"
     t.string "name", null: false
+    t.string "operational_state", default: "unknown", null: false
     t.text "prusalink_key"
     t.datetime "updated_at", null: false
     t.index ["name"], name: "index_printers_on_name", unique: true